refactor: modularize relationship constraints
